+++ irodsbackend/api.py
@@ -10,6 +10,7 @@ import string
 import uuid
 
 from contextlib import contextmanager
+from packaging import version
 
 from irods.api_number import api_number
 from irods.collection import iRODSCollection
@@ -500,9 +501,27 @@ class IrodsAPI:
         :param irods: iRODSSession object
         :return: String
         """
-        return '.'.join(
-            str(x) for x in irods.pool.get_connection().server_version
-        )
+        return '.'.join(str(x) for x in irods.server_version)
+
+    @classmethod
+    def get_access_lookup(cls, irods):
+        """
+        Return an ACL lookup dict compatible with the currently used iRODS
+        server version (4.2 and 4.3 supported).
+
+        :param irods: iRODSSession object
+        :return: Dict
+        """
+        v = version.parse(cls.get_version(irods))
+        d = '_' if v >= version.parse('4.3') else ' '
+        return {
+            'read': 'read{}object'.format(d),
+            'read{}object'.format(d): 'read',
+            'write': 'modify{}object'.format(d),
+            'modify{}object'.format(d): 'write',
+            'null': 'null',
+            'own': 'own',
+        }
 
     def get_object_stats(self, irods, path):
         """