+++ taskflowbackend/flows/landing_zone_move.py
@@ -41,6 +41,7 @@ class Flow(BaseLinearFlow):
         sample_path = self.irods_backend.get_path(zone.assay)
         zone_path = self.irods_backend.get_path(zone)
         admin_name = self.irods.username
+        access_lookup = self.irods_backend.get_access_lookup(self.irods)
 
         # HACK: Set zone status in the Django site
         zone.set_status(
@@ -196,6 +197,7 @@ class Flow(BaseLinearFlow):
                     'access_name': 'own',
                     'path': zone_path,
                     'user_name': admin_name,
+                    'access_lookup': access_lookup,
                     'irods_backend': self.irods_backend,
                 },
             )
@@ -210,6 +212,7 @@ class Flow(BaseLinearFlow):
                     'access_name': 'read',
                     'path': zone_path,
                     'user_name': zone.user.username,
+                    'access_lookup': access_lookup,
                     'irods_backend': self.irods_backend,
                 },
             )
@@ -226,6 +229,7 @@ class Flow(BaseLinearFlow):
                         'access_name': 'read',
                         'path': zone_path,
                         'user_name': self.flow_data['script_user'],
+                        'access_lookup': access_lookup,
                         'irods_backend': self.irods_backend,
                     },
                 )
@@ -293,6 +297,7 @@ class Flow(BaseLinearFlow):
                     'src_paths': zone_objects,
                     'access_name': 'read',
                     'user_name': project_group,
+                    'access_lookup': access_lookup,
                     'irods_backend': self.irods_backend,
                 },
             )
@@ -307,6 +312,7 @@ class Flow(BaseLinearFlow):
                     'access_name': 'null',
                     'path': sample_path,
                     'user_name': zone.user.username,
+                    'access_lookup': access_lookup,
                     'irods_backend': self.irods_backend,
                 },
             )
@@ -322,6 +328,7 @@ class Flow(BaseLinearFlow):
                         'access_name': 'null',
                         'path': sample_path,
                         'user_name': self.flow_data['script_user'],
+                        'access_lookup': access_lookup,
                         'irods_backend': self.irods_backend,
                     },
                 )